co-locate 'gql' typeDefs
