+++ src/resolvers/Tables/typeDefs.ts
@@ -1,0 +1,44 @@
+import { gql } from "apollo-server"
+
+export const dynamoTableTypeDef = gql`
+  type ProvisionedThroughput {
+    LastIncreaseDateTime: String
+    LastDecreaseDateTime: String
+    NumberOfDecreasesToday: Int
+    ReadCapacityUnits: Int
+    WriteCapacityUnits: Int
+  }
+
+  type KeySchema {
+    AttributeName: String
+    KeyType: String
+  }
+
+  type AttributeDefinitions {
+    AttributeName: String
+    AttributeType: String
+  }
+
+  type TableDescription {
+    TableName: String
+    TableStatus: String
+    CreationDateTime: String
+    TableSizeBytes: Int
+    ItemCount: Int
+    TableArn: String
+    ProvisionedThroughput: ProvisionedThroughput
+    KeySchema: [KeySchema]
+    AttributeDefinitions: [AttributeDefinitions]
+  }
+
+  type Table {
+    TableDescription: TableDescription
+  }
+
+  extend type Mutation {
+    createTable: Table @development @auth
+  }
+  extend type Query {
+    listTables: [String] @auth
+  }
+`
