+++ src/resolvers/Votes/typeDefs.ts
@@ -1,0 +1,22 @@
+import { gql } from "apollo-server"
+
+export const voteTypeDef = gql`
+  type Vote {
+    PK: ID
+    SK: String
+    createdAt: Date
+    updatedAt: Date
+    productName: String
+    rating: Float
+    username: String
+  }
+
+  extend type Mutation {
+    createVote(productName: String!, username: String!): Vote @auth
+  }
+
+  extend type Query {
+    queryVotesByEmail(email: String): [Vote]
+    queryVotesByProduct(productName: String): [Vote]
+  }
+`
